feat: invoice maintenance, service quality evaluation, collateral master v1.1, and webhook resilience
- Add Service Quality Evaluation feature (AppraisalEvaluation aggregate, CQRS, vw_AppraisalEvaluationList, migration AddAppraisalEvaluation)
- Add Invoice Maintenance feature (Invoice/InvoiceItem aggregate, Create/Update/Submit/Approve commands, list/detail/eligible-assignments queries, vw_InvoiceList, vw_EligibleAssignments, running number)
- Collateral Master v1.1: engagement-per-appraisal granularity, RowVersion concurrency token, field rename + dead-field drop, CollateralValueModel addition, multi-title support; 6 migrations; integration test suites PR4-PR8 + PhaseC_PR2
- Webhook delivery resilience: drop persisted retry state, add WebhookAttemptCounterHandler via Microsoft.Extensions.Http.Resilience; add collateral DLQ runbook
- Appraisal list/copy-template view updates and decision summary additions aligned with persisted Appraisal.Status
- Quotation cleanup: remove dead SelectQuotationFromIntegration command, add QuotationActor abstraction
- Auth menu seed entries for new features
- Add Database/publish/ to .gitignore

+++ Modules/Integration/Integration/Application/Features/AppraisalResults/GetAppraisalResult/GetAppraisalResultQuery.cs
@@ -41,7 +41,7 @@ public record AppraisalResultCollateral(
     string? HouseNo,
     string? BuildingType,
     int? BuildingAge,
-    int? TotalFloor,
+    decimal? TotalFloor,
     // Condo
     string? RoomNo,
     string? FloorNo,
